+++ beets/autotag/hooks.py
@@ -128,6 +128,7 @@ class TrackInfo(object):
     - ``index``: position on the entire release
     - ``medium``: the disc number this track appears on in the album
     - ``medium_index``: the track's position on the disc
+    - ``medium_total``: the number of tracks on the item's disc
     - ``artist_sort``: name of the track artist for sorting
     - ``disctitle``: name of the individual medium (subtitle)
     - ``artist_credit``: Recording-specific artist name
@@ -138,8 +139,8 @@ class TrackInfo(object):
     """
     def __init__(self, title, track_id, artist=None, artist_id=None,
                  length=None, index=None, medium=None, medium_index=None,
-                 artist_sort=None, disctitle=None, artist_credit=None,
-                 data_source=None, data_url=None):
+                 medium_total=None, artist_sort=None, disctitle=None,
+                 artist_credit=None, data_source=None, data_url=None):
         self.title = title
         self.track_id = track_id
         self.artist = artist
@@ -148,6 +149,7 @@ class TrackInfo(object):
         self.index = index
         self.medium = medium
         self.medium_index = medium_index
+        self.medium_total = medium_total
         self.artist_sort = artist_sort
         self.disctitle = disctitle
         self.artist_credit = artist_credit